@@ -2,25 +2,26 @@
     "name": "los/api-server",
     "description": "PHP api server middleware",
     "require": {
-        "php": "^7.2",
-        "psr/container": "^1.0",
+        "php": "^7.4 | ^8.0",
+        "ext-json": "*",
+        "psr/container": "^1.0 | ^2.0",
         "psr/http-message": "^1.0",
         "laminas/laminas-inputfilter": "^2.7",
         "laminas/laminas-paginator": "^2.7",
         "laminas/laminas-eventmanager": "^3.0",
         "psr/http-server-middleware": "^1.0",
-        "mezzio/mezzio-hal": "^1.0",
-        "mezzio/mezzio-helpers": "^5.0",
+        "mezzio/mezzio-hal": "^2.0",
+        "mezzio/mezzio-helpers": "^5.6",
         "mezzio/mezzio-problem-details": "^1.0",
         "laminas/laminas-diactoros": "^1.7 || ^2.2",
-        "los/uql": "^1.0"
+        "los/uql": "^1.1",
+        "doctrine/coding-standard": "^9.0"
     },
     "require-dev": {
-        "squizlabs/php_codesniffer": "^2.7",
+        "squizlabs/php_codesniffer": "^3.6",
         "phpstan/phpstan": "^0.12",
-        "laminas/laminas-coding-standard": "^1.0",
         "laminas/laminas-db": "^2.10",
-        "laminas/laminas-hydrator": "^3.0"
+        "laminas/laminas-hydrator": "^4.0"
     },
     "license": "MIT",
     "autoload-dev": {
